vin-detector/pyproject.toml
[tool.poetry]
name = "vin-detector"
version = "0.1.0"
description = "VIN Detector Service — klio-01"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.29"}
httpx = "^0.27"
aio-pika = "^9.4"
pydantic = "^2.7"
pydantic-settings = "^2.2"
python-multipart = "^0.0.9"
pillow = "^10.0"
ultralytics = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
